splitted some tests for readability, removed command line interactions for option parameters hence no more edit option
